LPS-196086 refactor InterestTermsChecker.java
moving class to InterestTermsMessageListener.java and deleting the other one
